backend/requirements.txt
aiohttp==3.8.0
aioredis==1.3.1
aiosignal==1.2.0
amqp==5.0.6
asgiref==3.4.1
astroid==2.8.5
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
billiard==3.6.4.0
bleach==4.1.0
Brotli==1.0.9
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
ConfigArgParse==1.5.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
Django==3.2.8
django-celery-beat==2.2.1
django-cleanup==5.2.0
django-common-helpers==0.9.2
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-developer-panel==0.1.2
django-environ==0.8.1
django-filter==21.1
django-redis==5.2.0
django-rest-swagger==2.2.0
django-silk==4.3.0
django-storages==1.12.3
django-timezone-field==4.2.1
django-xff==1.3.0
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
docutils==0.18
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
frozenlist==1.2.0
future==0.18.2
geocoder==1.38.1
geoip2==4.4.0
gevent==21.12.0
geventhttpclient==1.5.3
gprof2dot==2021.2.21
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
imageio==2.13.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.8.2
incremental==21.3.0
ip2geotools==0.1.6
IP2Location==8.6.4
isort==5.10.1
itsdangerous==2.1.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.0.2
jmespath==0.10.0
keyring==23.2.1
kombu==5.2.0
lazy-object-proxy==1.6.0
locust==2.8.4
lxml==4.6.4
MarkupSafe==2.0.1
maxminddb==2.2.0
mccabe==0.6.1
mongoengine==0.23.1
moviepy==1.0.3
msgpack==1.0.3
multidict==5.2.0
numpy==1.21.4
openapi-codec==1.3.2
outcome==1.1.0
packaging==21.2
Pillow==8.4.0
pip-review==1.1.0
pkginfo==1.7.1
platformdirs==2.4.0
polyglot==16.7.4
priority==1.3.0
proglog==0.1.9
prompt-toolkit==3.0.22
psutil==5.9.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld2==0.41
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.11.1
pymongo==3.12.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyquery==1.4.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
ratelim==0.1.6
readme-renderer==30.0
redis==3.5.3
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
roundrobin==0.0.2
s3transfer==0.5.0
SecretStorage==3.3.1
selenium==4.0.0
service-identity==21.1.0
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stripe==2.65.0
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
twine==3.6.0
Twisted==21.7.0
txaio==21.2.1
typed-ast==1.5.0
typing==3.7.4.3
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
